Additional checks if the grave fails to place #2
It will now check a 5x5x5 radius for the first possible valid position.

If this fails, I don't know what to tell you. That's a pretty large area for there to be absolutely no open spaces :P

If you want extra security, enable "forceGrave" in the config. This will overwrite the block you are standing in, so be careful.